Add files via upload
AttendEase is an automatic attendance monitoring system that uses face recognition technology. It is a computer-based system that automatically identifies and verifies a person’s identity based on their facial features. The system captures an image of a person’s face using a camera, compares it with a pre-registered image in a database, and
then records the attendance of the person.This system replaces the conventional attendance monitoring system which is inefficient and time-consuming. The system automates the attendance marking and periodically generates attendance reports that will be mailed to the respective faculties. 
AttendEase offers numerous benefits, such as reducing the risk of proxy attendance, improving security and accuracy, and reducing administrative workload. It is also a cost-effective solution that can save time, money, and resources in the long run. This system is implemented using OpenCV-Python.